+++ locale/opsiPackageBuilder_en.ts
@@ -886,102 +886,102 @@ The following message(s) returned:</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../oPB/core/processing.py" line="222"/>
+        <location filename="../oPB/core/processing.py" line="225"/>
         <source>Package file could not be found!</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../oPB/core/processing.py" line="1020"/>
+        <location filename="../oPB/core/processing.py" line="1038"/>
         <source>Error establishing SSH connection. See Log for details.</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../oPB/core/processing.py" line="1004"/>
+        <location filename="../oPB/core/processing.py" line="1022"/>
         <source>Command not found. See Log for details.</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../oPB/core/processing.py" line="1076"/>
+        <location filename="../oPB/core/processing.py" line="1094"/>
         <source>There are umlauts in some fields which can&apos;t be processed: build failed.</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../oPB/core/processing.py" line="1081"/>
+        <location filename="../oPB/core/processing.py" line="1099"/>
         <source>Backend error: installation failed.</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../oPB/core/processing.py" line="1086"/>
+        <location filename="../oPB/core/processing.py" line="1104"/>
         <source>Could not process &apos;install&apos; command.</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../oPB/core/processing.py" line="1091"/>
+        <location filename="../oPB/core/processing.py" line="1109"/>
         <source>Package file error. Check log.</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../oPB/core/processing.py" line="1141"/>
+        <location filename="../oPB/core/processing.py" line="1159"/>
         <source>Undefined error occurred. Check log.</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../oPB/core/processing.py" line="656"/>
+        <location filename="../oPB/core/processing.py" line="674"/>
         <source>Error during command execution. Check Log for details.</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../oPB/core/processing.py" line="346"/>
+        <location filename="../oPB/core/processing.py" line="361"/>
         <source>Copying file to:</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../oPB/core/processing.py" line="350"/>
+        <location filename="../oPB/core/processing.py" line="365"/>
         <source>Extracting...</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../oPB/core/processing.py" line="1096"/>
+        <location filename="../oPB/core/processing.py" line="1114"/>
         <source>Error during package extraction. Check log.</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../oPB/core/processing.py" line="1106"/>
+        <location filename="../oPB/core/processing.py" line="1124"/>
         <source>Permission denied. Check log.</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../oPB/core/processing.py" line="1111"/>
+        <location filename="../oPB/core/processing.py" line="1129"/>
         <source>Network logon failed. Check log.</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../oPB/core/processing.py" line="1116"/>
+        <location filename="../oPB/core/processing.py" line="1134"/>
         <source>Possible error: command or file not found. Check log.</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../oPB/core/processing.py" line="1121"/>
+        <location filename="../oPB/core/processing.py" line="1139"/>
         <source>Failed to get ip address for host. Check log.</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../oPB/core/processing.py" line="1131"/>
+        <location filename="../oPB/core/processing.py" line="1149"/>
         <source>The requested product action is not possible, because it is currently locked on the server. Check log.</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../oPB/core/processing.py" line="1126"/>
+        <location filename="../oPB/core/processing.py" line="1144"/>
         <source>Failed to execute opsi command. Check log.</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../oPB/core/processing.py" line="1101"/>
+        <location filename="../oPB/core/processing.py" line="1119"/>
         <source>Permission denied during opsi-makeproductfile. Check log or raise log level for opsi-makeproductfile.</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../oPB/core/processing.py" line="819"/>
+        <location filename="../oPB/core/processing.py" line="837"/>
         <source>opsi package update process already running.</source>
         <translation></translation>
     </message>
@@ -1025,7 +1025,7 @@ The following message(s) returned:</source>
 <context>
     <name>ProgressHook</name>
     <message>
-        <location filename="../oPB/core/processing.py" line="1318"/>
+        <location filename="../oPB/core/processing.py" line="1336"/>
         <source>In progress:</source>
         <translation></translation>
     </message>
@@ -1772,7 +1772,7 @@ The following message(s) returned:</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../oPB/controller/base.py" line="467"/>
+        <location filename="../oPB/controller/base.py" line="470"/>
         <source>Selection canceled! Using default opsi server from settings.</source>
         <translation></translation>
     </message>
@@ -1829,6 +1829,11 @@ The following message(s) returned:</source>
     <message>
         <location filename="../oPB/controller/base.py" line="425"/>
         <source>Run opsi package update process...</source>
+        <translation></translation>
+    </message>
+    <message>
+        <location filename="../oPB/controller/base.py" line="460"/>
+        <source>only config server - no repository usage</source>
         <translation></translation>
     </message>
 </context>
@@ -2758,17 +2763,17 @@ The following message(s) returned:</source>
 <context>
     <name>quickuninstallController</name>
     <message>
-        <location filename="../oPB/controller/components/scheduler.py" line="326"/>
+        <location filename="../oPB/controller/components/scheduler.py" line="376"/>
         <source>product id</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../oPB/controller/components/scheduler.py" line="326"/>
+        <location filename="../oPB/controller/components/scheduler.py" line="376"/>
         <source>version</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../oPB/controller/components/scheduler.py" line="326"/>
+        <location filename="../oPB/controller/components/scheduler.py" line="376"/>
         <source>description</source>
         <translation></translation>
     </message>
@@ -2786,27 +2791,27 @@ The following message(s) returned:</source>
 <context>
     <name>schedulerController</name>
     <message>
-        <location filename="../oPB/controller/components/scheduler.py" line="272"/>
+        <location filename="../oPB/controller/components/scheduler.py" line="322"/>
         <source>Do you really want to remove the selected job id(s)? This can&apos;t be undone!</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../oPB/controller/components/scheduler.py" line="285"/>
+        <location filename="../oPB/controller/components/scheduler.py" line="335"/>
         <source>Do you really want to remove all job(s)? This can&apos;t be undone!</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../oPB/controller/components/scheduler.py" line="298"/>
+        <location filename="../oPB/controller/components/scheduler.py" line="348"/>
         <source>No opsi client selected.</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../oPB/controller/components/scheduler.py" line="303"/>
+        <location filename="../oPB/controller/components/scheduler.py" line="353"/>
         <source>No opsi product selected.</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../oPB/controller/components/scheduler.py" line="306"/>
+        <location filename="../oPB/controller/components/scheduler.py" line="356"/>
         <source>Create AT jobs now?</source>
         <translation></translation>
     </message>
@@ -2824,12 +2829,12 @@ The following message(s) returned:</source>
 <context>
     <name>schedulerController_jobcreator</name>
     <message>
-        <location filename="../oPB/controller/components/scheduler.py" line="323"/>
+        <location filename="../oPB/controller/components/scheduler.py" line="373"/>
         <source>client</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../oPB/controller/components/scheduler.py" line="323"/>
+        <location filename="../oPB/controller/components/scheduler.py" line="373"/>
         <source>desc</source>
         <translation></translation>
     </message>
@@ -2837,37 +2842,37 @@ The following message(s) returned:</source>
 <context>
     <name>schedulerController_joblist</name>
     <message>
-        <location filename="../oPB/controller/components/scheduler.py" line="315"/>
+        <location filename="../oPB/controller/components/scheduler.py" line="365"/>
         <source>client</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../oPB/controller/components/scheduler.py" line="315"/>
+        <location filename="../oPB/controller/components/scheduler.py" line="365"/>
         <source>packet</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../oPB/controller/components/scheduler.py" line="315"/>
+        <location filename="../oPB/controller/components/scheduler.py" line="365"/>
         <source>action</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../oPB/controller/components/scheduler.py" line="315"/>
+        <location filename="../oPB/controller/components/scheduler.py" line="365"/>
         <source>date</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../oPB/controller/components/scheduler.py" line="315"/>
+        <location filename="../oPB/controller/components/scheduler.py" line="365"/>
         <source>time</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../oPB/controller/components/scheduler.py" line="315"/>
+        <location filename="../oPB/controller/components/scheduler.py" line="365"/>
         <source>AT jobid</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../oPB/controller/components/scheduler.py" line="315"/>
+        <location filename="../oPB/controller/components/scheduler.py" line="365"/>
         <source>user</source>
         <translation></translation>
     </message>